Mautic removes mautic:disable-tracking attribute for links in dwc
I was trying to use the mautic:disable-tracking="true" attribute for a specific link inside a Dynamic Content for Landing Pages. The attribute was removed every time exited the code editor.

The PR: https://github.com/mautic/mautic/pull/10355 unfortunately did not solved this issue also.
I do not really understand why is that so.

Anyway, I tested the present changes and the the mautic:disable-tracking="true" attribute is not removed anymore from a Dynamic Content for Landing Pages.